Shuffle around http/API code.
