+++ lib/API.h
@@ -1,15 +1,6 @@
 #ifndef __API_H
 #define __API_H
 
-#define SHA512_SZ 64
-
-#define RESPONSE_BAD_NONCE     5
-#define RESPONSE_ACCESS_DENIED 4
-#define RESPONSE_BAD_JSON      3
-#define RESPONSE_BAD_HTTP      2
-#define RESPONSE_BAD_CKSUM     1
-#define RESPONSE_GOOD          0
-
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -18,6 +9,10 @@ unsigned char val(char *i);
 unsigned char parse_response(char *in, struct cJSON **out, char *key, unsigned char key_len, char *rv, unsigned char rv_len);
 void get_hash(struct cJSON *data, char *sha_buf, char *key, unsigned char key_len);
 void print_hex(char *str, char *src, unsigned char sz);
+
+char *get_signed_packet(struct cJSON *data);
+unsigned char http_request(struct cJSON *data, struct cJSON **result, char *rand);
+void add_random_response(struct cJSON *data, char *rand);
 
 #ifdef __cplusplus
 }
