+++ lib/http.h
@@ -3,6 +3,15 @@
 
 #include "config.h"
 
+#define SHA512_SZ 64
+
+#define RESPONSE_BAD_NONCE     5
+#define RESPONSE_ACCESS_DENIED 4
+#define RESPONSE_BAD_JSON      3
+#define RESPONSE_BAD_HTTP      2
+#define RESPONSE_BAD_CKSUM     1
+#define RESPONSE_GOOD          0
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -12,9 +21,6 @@ void log_temp(unsigned long temp, char *name);
 void update_soda_status(unsigned char sold_out_mask);
 void update_nonce(void);
 unsigned char can_vend(unsigned long badge);
-char *get_signed_packet(struct cJSON *data);
-unsigned char http_request(struct cJSON *data, struct cJSON **result, char *rand);
-void add_random_response(struct cJSON *data, char *rand);
 
 #ifdef PLATFORM_ARDUINO
 #define RAND_SIZE (2 * sizeof(unsigned long))
